discord-api/requirements.txt
discord.py
transformers
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu
flair
lemminflect
